feat: add command for loading applications to device
This implements support for loading an application binary onto a TKey
device.

Besides the new client subcommand and corresponding command handler,
TKey instances now have a load_app method which accepts the path to a
binary file that will be uploaded in chunks of 127 bytes.

When the entire binary has been uploaded, the device will respond with a
BLAKE2s-256 hash digest for the data it has received. This digest will
be compared with the source binary to ensure all data has been received
intact.