feat: handle user-supplied secret when loading applications
This implements support for providing a User-Supplied Secret (USS) when
loading an application onto the device.

A command-line option has been added to the load subcommand. When this
option is set, the user will be asked for their secret without echoing
it to the terminal. The secret will then be sent along when starting the
application loading session.