feat: add support for a context manager
This implements the necessary methods for using the TKey class as a
context manager. It can be used to ensure that the serial connection is
opened before and closed after interacting with the TKey device.

The example client handlers has been rewritten to use the context
manager instead of using the connect/disconnect methods on the
instantiated TKey object.

Information on how to use the context manager has been added to the
README.