feat: add command for device info
Add support for fetching firmware name and version from a TKey device.

Besides the additional client subcommand and corresponding client
handler, it includes the necessary helper functions to deal with and
debug the TKey framing, command and response protocol, as well as
writing data to and reading data from the TKey device.

The protocol implementation currently does not ensure that the response
frames correspond to the specific command sent to the device, this
should be considered as an improvement in the future.